+++ 1.2-pong/pong5/hitbox.h
@@ -13,10 +13,10 @@
 
 struct Hitbox
 {
-    int x1;
-    int y1;
-    int x2;
-    int y2;
+    float x1;
+    float y1;
+    float x2;
+    float y2;
 };
 
 int collides(struct Hitbox hitbox1, struct Hitbox hitbox2);
